Plugin mock needs clarification
If we mock a plugin run so that it belongs to janedoe
I get a duplicate entry error when I log on. This might be due to
mocking. Or due to the fact that the eval system is uing pwd and
the web openid-connect. This would be bad. In that case we would
have to use openid-connect in the eval sysetem. Which might be fun.